# packages/api/openapi.v1.json
{
  "components": {
    "schemas": {
      "ApiEnvelope_ArtifactRef_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ArtifactRef"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiEnvelope[ArtifactRef]",
        "type": "object"
      },
      "ApiEnvelope_ComparisonResult_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ComparisonResult"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiEnvelope[ComparisonResult]",
        "type": "object"
      },
...
      "ApiEnvelope_LanguageProfile_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/LanguageProfile"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiEnvelope[LanguageProfile]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Published projections and authenticated, review-gated submission requests.",
    "title": "PolyCodeBench Public API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/artifacts/{artifact_id}": {
      "get": {
        "description": "Mint a short-lived download URL only for an artifact in a published release.",
        "operationId": "get_artifact_download_link_v1_artifacts__artifact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "release",
            "required": true,
            "schema": {
              "maxLength": 120,
              "minLength": 1,
              "title": "Release",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_ArtifactRef_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Artifact Download Link"
      }
    },
    "/v1/artifacts/{artifact_id}/download": {
      "get": {
        "description": "Serve one bounded, digest-checked object after release and token revalidation.",
        "operationId": "download_public_artifact_v1_artifacts__artifact_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "release",
            "required": true,
            "schema": {
              "maxLength": 120,
              "minLength": 1,
              "title": "Release",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "token",
            "required": true,
            "schema": {
              "maxLength": 512,
              "minLength": 32,
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Public Artifact"
      }
    },
    "/v1/compare": {
      "get": {
        "operationId": "get_compare_v1_compare_get",
        "parameters": [
          {
            "in": "query",
            "name": "models",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "title": "Models",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "language",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "title": "Language",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "family",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "title": "Family",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "difficulty",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "title": "Difficulty",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "task_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task Date"
            }
          },
          {
            "in": "query",
            "name": "release",
            "required": false,
            "schema": {
              "default": "latest",
              "title": "Release",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_ComparisonResult_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Compare"
      }
    },
    "/v1/languages/{language_id}": {
      "get": {
        "operationId": "get_language_v1_languages__language_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "language_id",
            "required": true,
            "schema": {
              "title": "Language Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "release",
            "required": false,
            "schema": {
              "default": "latest",
              "title": "Release",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_LanguageProfile_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Language"
      }
    },
    "/v1/leaderboard": {
      "get": {
        "operationId": "get_leaderboard_v1_leaderboard_get",
        "parameters": [
          {
            "in": "query",
            "name": "language",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "title": "Language",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "family",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "title": "Family",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "difficulty",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "title": "Difficulty",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "task_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task Date"
            }
          },
          {
            "in": "query",
            "name": "release",
            "required": false,
            "schema": {
              "default": "latest",
              "title": "Release",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "rank_asc",
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_list_LeaderboardEntry__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Leaderboard"
      }
    },
    "/v1/methodology/{version}": {
      "get": {
        "description": "The frozen methods behind a release, resolved by version as release links expect.",
        "operationId": "get_methodology_v1_methodology__version__get",
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "release",
            "required": false,
            "schema": {
              "default": "latest",
              "title": "Release",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_Methodology_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Methodology"
      }
    },
...
    "/v1/models/{model_config_id}": {
      "get": {
        "operationId": "get_model_v1_models__model_config_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "model_config_id",
            "required": true,
            "schema": {
              "title": "Model Config Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "release",
            "required": false,
            "schema": {
              "default": "latest",
              "title": "Release",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_ModelProfile_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Model"
      }
    },
...
    "/v1/releases/{release_id}": {
      "get": {
        "operationId": "get_release_v1_releases__release_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "release_id",
            "required": true,
            "schema": {
              "title": "Release Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_ReleaseSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Release"
      }
    },
...
    "/v1/tasks/{task_id}": {
      "get": {
        "operationId": "get_task_v1_tasks__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "release",
            "required": false,
            "schema": {
              "default": "latest",
              "title": "Release",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_TaskSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Task"
      }
    },
...
  }
}